+++ src/App.jsx
@@ -12,23 +12,28 @@ function fetchTodos() {
   return result;
 }
 
-function App() {
-  const [todos, setTodos] = useState(fetchTodos());
-  useEffect(() => {
-    const storedTodos = fetchTodos();
-    setTodos(storedTodos);
-}, []);
- 
-
- 
+  function App() {
+    const [todos, setTodos] = useState(fetchTodos());
+    useEffect(() => {
+      const storedTodos = fetchTodos();
+      setTodos(storedTodos);
+  }, []);
   
-
-
+  const handleRemove = (todo, index) => {
+    const result = todos.filter(todoItem => {
+        if (todoItem !== todo) {
+        return true;
+        }
+    })
+    setTodos(result);
+    localStorage.removeItem(result);
+  }
+ 
   return (
     <div>
       <TodoHeader />
       <TodoInput />
-      <TodoList todos={todos} />
+      <TodoList todos={todos} onTodoRemove={handleRemove} />
     </div>
   )
 }
